requirements.txt
tensorflow
codecarbon
pandas
datetime
time
os
matplotlib.pyplot
csv
PIL